requirements.txt
annotated-types==0.5.0
asgiref==3.7.2
astroid==2.15.6
beautifulsoup4==4.12.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
crispy-bootstrap4==2022.1
cryptography==38.0.1
defusedxml==0.7.1
dill==0.3.6
dj-database-url==0.5.0
Django==4.1.2
django-auditlog==2.3.0
django-bootstrap-datepicker-plus==5.0.4
django-bootstrap4==23.2
django-cleanup==5.2.0
django-crispy-forms==2.0
django-filter==23.2
django-formtools==2.4.1
django-heroku==0.3.1
django-jazzmin==2.4.8
django-mathfilters==1.0.0
django-otp==1.2.2
django-phonenumber-field==6.4.0
django-raw-query==0.0.4
django-two-factor-auth==1.15.4
django-widget-tweaks==1.4.8
gunicorn==20.1.0
idna==3.4
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
mysql==0.0.3
mysql-connector-python==8.0.33
mysqlclient==2.2.0
numpy==1.25.0
oauthlib==3.2.2
opencv-python==4.8.0.74
Pillow==9.2.0
platformdirs==3.9.1
protobuf==3.20.3
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==2.0.3
pydantic_core==2.3.0
PyJWT==2.6.0
pylint==2.17.4
pypng==0.20220715.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2022.5
qrcode==7.4.2
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
soupsieve==2.4.1
sqlparse==0.4.3
toml==0.10.2
tomlkit==0.11.8
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.12
whitenoise==6.5.0
wrapt==1.15.0
